[Elasticsearch] Move implementation to own folder
